Refactor split-bill file structure to best practices
I.e. dedicated folder for Split, and ErrorBubble
